+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Version 2.2.3 (?)
 
+* [new] Add support for HTML5 mode (pretty urls)
 * [new] Add the ability to specify an `optional` attribute on any fragment configuration, allowing application to load anyway.
 * [new] Add the ability to specify an `ignore` attribute on any fragment configuration to avoid loading it.